Dynamically import all modules in ext
This is a pretty hackish solution, but it works. It would be nice if Python had
a built in way to do this. __all__ doesn't work because you'd have to list off
every extension in it and that's not what we're looking for.

Credit to Lennard Regebro on Stack Exchange