Very Basic Database Module
Only supports MySQL. It uses a PDO-style DSN (I had a regex for it handy). DSN
needs to be stored in core/Config.py, so this introduces a need for an install
script. Perhaps there's a better way to store this config info (e.g just read
it in from a plain text file).