Chdir to Location of main.py
The working directory needs to be tho location of main.py so that CONFIG can be
loaded. While working on this, I noticed that just how much persistence there
is between requests, so it may be possible to do all the initialization
statically, rather than doing it on every request.