Cleanup nonstatic extensions at the end of each request
Before, they would be cleaned up whenever the garbage collector got around to
it, which is usually when the interpreters were being shutdown. This makes sure
the deconstructors are consistently called at the end of each request. One
thing to note however is that the deconstructors for static extensions will
probably never be called, so don't count on it.